+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### Added
+- ReportingApi with attachment support
+- `extendTestDescriptionWithLastError` option to the RP config to be able to toggle the last error log attaching to the test description.
+### Changed
+- `@reportportal/client-javascript` bumped to version `5.1.3`.
 
 ## [5.0.0] - 2024-02-15
 ### Added